добавлены axios запросы, изменены запросы на питоне
