+++ python-backend/api/HotelBooking.py
@@ -1,13 +1,13 @@
-from flask import Blueprint, jsonify
-import pymongo
-from config import DevelopmentConfig
 import os
-from dotenv import load_dotenv
+from datetime import datetime
+import pymongo
 from bson.objectid import ObjectId
+from dotenv import load_dotenv
+from flask import Blueprint, jsonify
 from flask_jwt_extended import jwt_required, get_jwt_identity
 
 hotel_booking_blueprint = Blueprint('hotel_booking_blueprint', __name__,
-                        template_folder='templates')
+                                    template_folder='templates')
 
 load_dotenv()
 print(os.getenv('MONGO_DB_URL'))
@@ -16,17 +16,81 @@ client = pymongo.MongoClient(os.getenv('MONGO_DB_URL'))
 db = client[os.getenv('MONGO_DOCUMENT')]
 
 
-@hotel_booking_blueprint.route('/hotelbooking/<booking_id>/cancel', methods=['POST'])
+@hotel_booking_blueprint.route('/hotelbooking/<booking_id>/pay', methods=['POST'])
+@jwt_required()  # Требуется JWT
+def pay_for_booking(booking_id):
+    try:
+        # Предполагается, что user_id получен из JWT или другого источника
+        user_id = get_jwt_identity()
+        print(user_id)
+
+        # Находим пользователя
+        user = db['users'].find_one({"_id": ObjectId(user_id)})
+        if not user:
+            return jsonify({"error": "User not found"}), 404
+        # НУЖНО ПРОВЕРИТЬ ЧТО БРОНЬ ПРИНАДЛЕЖИТ ЕМУ
+
+        # Поиск брони по идентификатору
+        booking = db["hotelbookings"].find_one({"_id": ObjectId(booking_id)})
+        if not booking:
+            return jsonify({"error": "HotelBooking not found"}), 404
+
+        # Поиск комнаты, чтобы узнать ее стоимость
+        room = db["rooms"].find_one({"_id": booking["room"]})
+        if not room:
+            return jsonify({"error": "Room not found"}), 404
+
+        # Поиск банковского аккаунта пользователя
+        account = db["bankaccount"].find_one({"_id": user["account"]})
+        if not account:
+            return jsonify({"error": "BankAccount not found"}), 404
+
+        # Проверка, достаточно ли денег на счету для оплаты
+        if int(account["amount"]["$numberLong"]) < room["price"]:
+            return jsonify({"error": "Insufficient funds"}), 400
+
+        # Создание транзакции
+        transaction = {
+            "user": user["_id"],
+            "date": datetime.now(),
+            "amount": -room["price"],  # Сумма списания
+            "description": f"Payment for booking {booking_id}"
+        }
+        db["accounttransaction"].insert_one(transaction)
+
+        # Обновление счета пользователя
+        new_amount = int(account["amount"]["$numberLong"]) - room["price"]
+        db["bankaccount"].update_one(
+            {"_id": account["_id"]},
+            {"$set": {"amount": {"$numberLong": str(new_amount)}}}
+        )
+
+        # Обновление статуса брони на "paid"
+        db["hotelbookings"].update_one(
+            {"_id": ObjectId(booking_id)},
+            {"$set": {"status": "Paid", "fixedPrice": room["price"]}}
+        )
+
+        # return jsonify({"message": "Booking successfully paid"}), 200
+        return jsonify(transaction), 200
+    except Exception as e:
+        # Любые другие ошибки отлавливаются здесь и возвращается ошибка 500
+        return jsonify({"error": f"Internal server error: {e}"}), 500
+
+
+@hotel_booking_blueprint.route('/hotelbooking/<booking_id>/status/cancel', methods=['POST'])
 @jwt_required()  # Требуется JWT
 def cancel_reservation(booking_id):
     try:
         # Предполагается, что user_id получен из JWT или другого источника
         user_id = get_jwt_identity()
         print(user_id)
+
         # Находим пользователя и проверяем, включено ли бронирование в его историю
         user = db['users'].find_one({"_id": ObjectId(user_id)})
         if not user:
-            return jsonify({"error": "Отмена доступа"}), 403
+            return jsonify({"error": "User not found"}), 404
+
         # booking_id_str = str(booking_id)  # Преобразуем ObjectId в строку
         # if not user or booking_id_str not in [h['$oid'] for h in user.get('hotelHistory', [])]:
         #     return jsonify({"error": "No rights to cancel this booking"}), 403
@@ -34,7 +98,7 @@ def cancel_reservation(booking_id):
         # Находим запись бронирования
         hotel_book = db['hotelbookings'].find_one({"_id": ObjectId(booking_id)})
         if not hotel_book:
-            return jsonify({"error": "Бронирование не найдено"}), 404
+            return jsonify({"error": "HotelBooking not found"}), 404
 
         # Обновляем статус бронирования на "cancelled"
         result = db['hotelbookings'].update_one(
@@ -44,13 +108,9 @@ def cancel_reservation(booking_id):
 
         # Проверяем, что обновление прошло успешно
         if result.matched_count:
-            return jsonify({"message": "Бронирование отменено"}), 200
+            return jsonify(result), 200
         else:
-            return jsonify({"error": "Не удалось обновить бронирование"}), 500
-
-    except pymongo.errors.PyMongoError as e:
-        # Обработка любых ошибок связанных с MongoDB
-        return jsonify({"error": f"Ошибка базы данных: {e}"}), 500
+            return jsonify({"error": "HotelBooking not updated"}), 500
     except Exception as e:
         # Обработка прочих ошибок
-        return jsonify({"error": f"Внутренняя ошибка сервера: {e}"}), 500
+        return jsonify({"error": f"Internal server error: {e}"}), 500
