+++ python-backend/app/__init__.py
@@ -1,19 +1,17 @@
-from flask import Flask
+import os
 import pymongo
 from dotenv import load_dotenv
-import os
+from flask import Flask
 from flask_jwt_extended import JWTManager
 from api.HotelBooking import hotel_booking_blueprint
-from api.Transaction import transaction_blueprint
-from api.User import user_api_blueprint
+from api.Profile import user_api_blueprint
 
 load_dotenv()
 
 app = Flask(__name__)
-# app.config.from_object("config.DevelopmentConfig")
+app.config.from_object("config.DevelopmentConfig")
 
 app.register_blueprint(hotel_booking_blueprint)
-app.register_blueprint(transaction_blueprint)
 app.register_blueprint(user_api_blueprint)
 # print(os.getenv('MONGO_DB_URL'))
 # client = pymongo.MongoClient(app.config['MONGO_DB_URL'])
